requirements.txt
Django==4.1.2
djangorestframework==3.14.0
frozenlist==1.3.1
future==0.18.2
idna==3.4
multidict==6.0.2
pefile==2022.5.30
Pillow==9.3.0
psycopg2==2.9.5
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.9
pytz==2022.2.1
pywin32-ctypes==0.2.0
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
yarl==1.8.1
